+++ toys-core/test/test_arg_parser.rb
@@ -678,7 +678,7 @@ describe Toys::ArgParser do
     it "errors on runnable tool if there are too many arguments" do
       tool.add_optional_arg(:b)
       tool.add_required_arg(:a)
-      tool.runnable = proc {}
+      tool.run_handler = proc {}
       arg_parser.parse(["foo", "bar", "baz"])
       arg_parser.finish
       assert_errors_include('Extra arguments: "baz".', arg_parser.errors)